address PR review comments
