+++ .changeset/serious-poems-retire.md
@@ -2,4 +2,4 @@
 "@clerk/clerk-js": patch
 ---
 
-Chore: Replace beforeEmit with an explicity call after `setActive`, inside the experimental UserVerification.
+Chore: Replace beforeEmit with an explicit call after `setActive`, inside the experimental UserVerification.
