+++ integration/templates/next-app-router/src/app/(reverification)/requires-re-verification/page.tsx
@@ -2,7 +2,6 @@ import { logUserIdActionReverification } from '../actions';
 import { ButtonAction } from '../button-action';
 
 function Page() {
-  // @ts-ignore
   return <ButtonAction action={logUserIdActionReverification} />;
 }
